test: remove duplicate unix socket tests
